refactor: abstract `FileSystem.cpExternal()`
